move viewer app to separate file
