+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Add PDBj as a pdb-provider option
+- Move Viewer APP to a separate file to allow use without importing light theme & index.html
 
 ## [v3.0.0-dev.8] - 2021-12-31
